requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3

# Database
PyMySQL==1.1.0
cryptography==41.0.4

# Security
Werkzeug==2.3.7

# Environment Management
python-dotenv==1.0.0

# Testing Dependencies
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2

# Production Server (optional)
gunicorn==21.2.0

# Utilities
click==8.1.7